refactor: use contains instead of index + separate code in another file
